+++ packages/backend/src/server.ts
@@ -1,13 +1,12 @@
 import express from 'express';
 import http from 'http';
 import cors from 'cors';
-// import { initializeWebSocket } from './websocket';
-import { initializeLLMConfig } from './lib/modelProviderService';
+import { initializeLLMConfig } from './config/llm';
 import { getPort } from '@starknet-agent/agents/config';
 import logger from './utils/logger';
-import { initializeHTTP } from './http';
-import { Container } from './types/context';
-import { validateConfig } from './config/server';
+import { initializeHTTP } from './config/http';
+import { Container } from './config/context';
+import { validateConfig } from './utils/validateConfig';
 
 export async function createApplication() {
   try {
@@ -42,11 +41,9 @@ export async function createApplication() {
 
     // Initialize both HTTP and WebSocket with container
     initializeHTTP(app, container);
-    // const wss = initializeWebSocket(server, container);
-    const wss = undefined;
 
     // Update container with initialized services
-    container.setContext({ app, wss });
+    container.setContext({ app });
 
     return { server, container };
   } catch (error) {
